Create submodules.txt
<repository_url> <branch_name> <folder1> <folder2> ... <folderN>

repository_url: The URL of the Git repository to clone.

branch_name: The specific branch to check out in the repository.

folder(s): The folder(s) within the repository to be included in the sparse-checkout operation. Multiple folders can be specified for each repository, and only these folders will be checked out.